Fixed interpolator default values
Closes #28

Commenting out ETrF min/max limits and setting etrf_ndvi_flag to False.  These values should be explicitly set by the user so that they are not suprised the first time they run the scripts.

Also, updated the file paths in the template INIs to better match the example (but they will be overwritten anyway).