Switched back to correct/default WRS2 shapefile naming
Since the current branch will already be somewhat incompatible with older versions (0.1.X), I switched back to using the default WRS2 shapefile naming convention (with "WRS" in uppercase).  I changed the example and template INIs but the user will need to update their own INIs also.  To help mitigate this I added additional code that will try and read the lower case version if the upper case is not found.  Since windows is fairly case insensitive with file names, I think this will only be an issue on Mac and Linux so at some point I should probably test that.