Fixed bug in LANDFIRE download script with older versions
Also, added code to standardize the naming of the "Grid" folder.  This folder is named Grid, grid, grid1, grid2, and Grid2 depending on the LANDFIRE version.  It might be better to leave the naming exactly as it is in the zip file but I think it will make it less confusing to the user if they try and switch LANDFIRE versions.