Addressed issue when nlcd_flag = False in project INI
Currently METRIC can't really run if the NLCD flag is set to false, since much of model2 is hardcoded to work with NLCD.  A stronger warning was added to let the user know this may be a problem and the prep_ini code was updated to populate an empty landuse_raster value through the INIs to make this more obvious.  The only reason to let the scripts even attempt to be run with nlcd_flag = False is that the user may not want to run model 2.

The correct fix for this though would be to try and support other landuse rasters.